moved the remove ncbi accession lookup into new subcommand taxit entrez
